requirements.txt
strands-agents[anthropic,otel]
strands-agents-tools
